refactor(history): Move evaluate_version_bump
