+++ tests/test_history.py
@@ -18,21 +18,21 @@ MAJOR_LAST_RELEASE_MINOR_AFTER = [MINOR, '1.1.0', MAJOR]
 
 class EvaluateVersionBumpTest(TestCase):
     def test_major(self):
-        with mock.patch('semantic_release.history.get_commit_log',
+        with mock.patch('semantic_release.history.logs.get_commit_log',
                         lambda: ALL_KINDS_OF_COMMIT_MESSAGES):
             self.assertEqual(evaluate_version_bump('0.0.0'), 'major')
 
     def test_minor(self):
-        with mock.patch('semantic_release.history.get_commit_log',
+        with mock.patch('semantic_release.history.logs.get_commit_log',
                         lambda: MINOR_AND_PATCH_COMMIT_MESSAGES):
             self.assertEqual(evaluate_version_bump('0.0.0'), 'minor')
 
     def test_patch(self):
-        with mock.patch('semantic_release.history.get_commit_log', lambda: PATCH_COMMIT_MESSAGES):
+        with mock.patch('semantic_release.history.logs.get_commit_log', lambda: PATCH_COMMIT_MESSAGES):
             self.assertEqual(evaluate_version_bump('0.0.0'), 'patch')
 
     def test_nothing_if_no_tag(self):
-        with mock.patch('semantic_release.history.get_commit_log', lambda: ['', '...']):
+        with mock.patch('semantic_release.history.logs.get_commit_log', lambda: ['', '...']):
             self.assertIsNone(evaluate_version_bump('0.0.0'))
 
     def test_force(self):
@@ -41,21 +41,21 @@ class EvaluateVersionBumpTest(TestCase):
         self.assertEqual(evaluate_version_bump('0.0.0', 'patch'), 'patch')
 
     def test_should_account_for_commits_earlier_than_last_commit(self):
-        with mock.patch('semantic_release.history.get_commit_log',
+        with mock.patch('semantic_release.history.logs.get_commit_log',
                         lambda: MAJOR_LAST_RELEASE_MINOR_AFTER):
             self.assertEqual(evaluate_version_bump('1.1.0'), 'minor')
 
     @mock.patch('semantic_release.history.config.getboolean', lambda *x: True)
-    @mock.patch('semantic_release.history.get_commit_log', lambda: [NO_TAG])
+    @mock.patch('semantic_release.history.logs.get_commit_log', lambda: [NO_TAG])
     def test_should_patch_without_tagged_commits(self):
         self.assertEqual(evaluate_version_bump('1.1.0'), 'patch')
 
     @mock.patch('semantic_release.history.config.getboolean', lambda *x: False)
-    @mock.patch('semantic_release.history.get_commit_log', lambda: [NO_TAG])
+    @mock.patch('semantic_release.history.logs.get_commit_log', lambda: [NO_TAG])
     def test_should_return_none_without_tagged_commits(self):
         self.assertIsNone(evaluate_version_bump('1.1.0'))
 
-    @mock.patch('semantic_release.history.get_commit_log', lambda: [])
+    @mock.patch('semantic_release.history.logs.get_commit_log', lambda: [])
     def test_should_return_none_without_commits(self):
         """
         Make sure that we do not release if there are no commits since last release.
